requirements.txt
langchain
langchain-openai
pydantic
chromadb
python-dotenv
openai